Switch back to 16 bit bias for speed

+++ src/nnue/nnue_accumulator.h
@@ -30,7 +30,7 @@ namespace Eval::NNUE {
 
   // Class that holds the result of affine transformation of input features
   struct alignas(kCacheLineSize) Accumulator {
-    std::int32_t
+    std::int16_t
         accumulation[2][kRefreshTriggers.size()][kTransformedFeatureDimensions];
     AccumulatorState state[2];
   };
